// src/models/Study.ts
export default interface Study {
  studyStart: string | null;
  totalTimeSpan: string | null;
  studyType: string | null;
  studyStatus: string | null;
  name: string | null;
  recordedTimeSpan: string | null;
  identityString: string | null;
  dateOfBirth: string | null;
  ageAtStudy: string | null;
  indicationForEeg: string | null;
  clinicalDiagnosis: string | null;
  diagnosticSignificance: string | null;
  clinicalComments: string | null;
  summary: string | null;
  lastEpisode: string | null;
  episodeFrequency: string | null;
  medication: string | null;
  brainMriCt: string | null;
  functionalNeuroimaging: string | null;
  medicationAdministeredDuringTheRecording: string | null;
  medicationWithdrawal: string | null;
  latestMeal: string | null;
  alertness: string | null;
  studyEnd: string | null;
  externalDatabase: string | null;
  postmenstrualAge: string | null;
  correctedAgeAtStudy: string | null;
  postMenstrualAgeAtStudy: string | null;
  externalStudyId: string | null;
  studyName: string | null;
  scalfDefects: string | null;
  studyNotes: string | null;
}